refactor: 合流の判断ロジックを MergeCoordinator へ切り出す (#108)
